refactor(post): extract PostList feed from the index page
Move the paginated post feed, sort watch and delete dialog into a
PostList component so index.vue is a bill of materials.
