+++ packages/app/app/pages/index.vue
@@ -1,41 +1,6 @@
-<script setup lang="ts">
-import { usePostStore } from "@/store/post";
-
-const { readMorePosts, readPosts } = useReadPosts();
-const { refresh } = await readPosts();
-const postStore = usePostStore();
-const { resetCursorPaginationData } = postStore;
-const { hasMore, items, sortType } = storeToRefs(postStore);
-
-watch(sortType, async () => {
-  resetCursorPaginationData();
-  await refresh();
-});
-</script>
-
 <template>
   <NuxtLayout>
-    <v-pull-to-refresh
-      min-h-16
-      @load="
-        async ({ done }) => {
-          await refresh();
-          done();
-        }
-      "
-    >
-      <v-container>
-        <PostSortMenu />
-        <v-divider my-2 />
-        <v-row>
-          <v-col v-for="post of items" :key="post.id" cols="12">
-            <PostCard :post />
-          </v-col>
-        </v-row>
-        <StyledWaypoint flex justify-center :is-active="hasMore" @change="readMorePosts" />
-      </v-container>
-    </v-pull-to-refresh>
-    <PostConfirmDeleteDialog />
+    <PostList />
     <template #left>
       <AppProductList />
     </template>
